Exploit patching and bug fixes (#886)
* Better item and stat limit checking

* Various skill fixes

* Kill time null check

* Timer expiration issues

* Correcting boosted base stats for mitama set bonuses

* More null checks

* Ignore correct tbl calculated stat adjustments that result in 0%

* Correcting digitalize times based on mitama demon type

* Flooring tech damage that hits to 1

* Fixing spirit defuse cost calculations

* Honoring demon locking for trifusion

* Implementing (somehow missed) macca and mag rates

* Loosening object interaction restrictions, at least temporarily

+++ server/channel/src/CharacterState.cpp
@@ -330,7 +330,8 @@ std::shared_ptr<objects::DigitalizeState> CharacterState::Digitalize(
         for(uint8_t i = (size_t)CorrectTbl::RES_DEFAULT;
             i <= (uint8_t)CorrectTbl::NRA_MAGIC; i++)
         {
-            mDigitalizeState->SetCorrectValues(i, demonStats[(CorrectTbl)i]);
+            mDigitalizeState->SetCorrectValues(i,
+                (int16_t)demonStats[(CorrectTbl)i]);
         }
     }
 
@@ -935,7 +936,7 @@ uint8_t CharacterState::RecalculateStats(
     {
         // Combat run speed can change from unadjusted stats (nothing
         // natively does this)
-        SetCombatRunSpeed(stats[CorrectTbl::MOVE2]);
+        SetCombatRunSpeed((int16_t)stats[CorrectTbl::MOVE2]);
 
         if(!mInitialCalc)
         {
@@ -1023,7 +1024,8 @@ uint8_t CharacterState::RecalculateStats(
     {
         for(auto statPair : stats)
         {
-            calcState->SetCorrectTbl((size_t)statPair.first, statPair.second);
+            calcState->SetCorrectTbl((size_t)statPair.first,
+                (int16_t)statPair.second);
         }
 
         return result;
